Fix the circular import
